Got rid of unused alt-h/alt-l mappings.
I put these in when I was doing a lot of editing of racket files for CS
360 so that I could do Dr. Racket alt-left/alt-right like behavior with
vim motions (jumping to the matching parenthesis in that direction).
Since vim already has the % command, which, although is perhaps more
cumbersome to use, works just fine doing essentially the same thing, I
decided it's time to remove these.