Removed unused DiffOrig command.
This was something I thought might be useful. It gives you a vimdiff of
the changes you've made to the file since it was loaded/written. I end
up saving my files a lot anyway, and rely on source control to show me
the diffs of the changes I've made since the last commit, and so this is
no longer necessary for me.