Restructure repo so that everything is inside a .vim or vimfiles folder.
Up to this point I had this repo set up so that its root was the user's
home directory, since I thought that that's where the .vimrc and/or
_vimrc file needed to be. I just found out that vim will find the files
if they're inside of a .vim directory on unix-based systems or a vimfiles
directory on Windows, which simplifies things so that I don't have to
have a weird .gitignore at the root that ignores everything except the
files I want tracked in the home directory, and I can just clone the
repo directly into a folder of the appropriate name on the given system
I'm cloning it to (vimfiles for Windows and .vim elsewhere). This also
meant that I can get rid of the symlinks and just have the vimrc file
named the same for both systems.