Update the backup, swap, and undo files to go into separate folders.
I had the swap and backup directories set to $HOME/vimfiles/tmp, but
with me changing it to get rid of the symlinks, that caused it to not be
able to create swap or backup files since it couldn't find the
non-existent directory (in WSL at least; in Windows proper it would have
been fine). I figured out how to reference the .vim or vimfiles
directory regardless of its name so that I could set the directory to a
subfolder of the .vim directory. I also added setting of the undo dir,
and split these into three different hidden directories.
Also removed the old date from the vimrc file since I've updated it a
bunch since then.