Move taglist plugin to not be loaded by default.
The plugin reports an error if ctags isn't installed, which is annoying
when starting vim. Moving it from pack/start to pack/opt makes it so
it's not loaded by default. If I decide I want to I can add some logic
to my vimrc that checks for ctags and calls "packadd! taglist" for me,
so that it's loaded by default when ctags is installed, but there isn't
an error if it's not. I think I previously had commented out the error
to get around this with the old plugin version, but this seems like a
better solution.