Configure tab completion in a popup window for command-line mode.
The `pum` option for the `wildoptions` setting isn't supported in older
versions of Vim, so I added a version check around setting that so as to
not break the config when used with older Vim installations. In that
case the regular wild menu will be used instead of the popup menu.