Add first version of Windows setup scripts.
Instead of manually creating symlinks every time I clone the repo and
set it up on a new machine, I'm making some scripts to help with the
setup. For right now it's just the vimfiles folder that needs to be
symlinked (and I'm also symlinking it as a .vim folder as well so that
vim inside git bash sees it). In the future I'll update these scripts to
make symlinks for other things that I decide to include in my dotfiles
repo, as well as make a Linux version of the setup scripts. I've got
some ideas for how to make this all a bit cleaner with some config files
that define the mappings to use for the symlinks, but this initial
version of these scripts works for now.