Switch toggles from F keys to leader key mappings.
These mappings will be easier to remember and to type than using F2,
F4/Shift-F4, F6/Shift-F6, and F8.
For right now I'm just using the default leader key of "\". I played
around with using space, but I'm not sure I like it as much given that I
currently use it for clearing highlights and messages.