Restructure to be dotfiles repo: move vim config under vimfiles subfolder.
This repo has up to this point been just for my vim config files, but I
want to repurpose it to hold all my dotfiles going forward. This change
is the first step in restructuring the repo. I've renamed the repo
folder to `.dotfiles` (from `vimfiles`) and I'm moving all the files
inside of a vimfiles subfolder.