Chapter 12 Exercises. Also sourced defaults.vim in vimrc.
I discovered that a lot of the default options that I wanted to have set
were normally set in defaults.vim, which would have been loaded if I
didn't have my own vimrc file. So I added a line at the top of my vimrc
to source defaults.vim to have the same baseline to start with. Neovim
of course didn't need that since it just has sensible defaults out of
the box.