Switch to using `<Cmd>` in mappings where possible.
This allowed me to combine the normal and insert mode mappings for
moving lines up and down. Some mappings in visual mode, etc., need to
continue using a regular `:` because they rely on the visual mode marks
getting added automatically to the beginning of the command as the range
to operate on, or they're more complicated commands including embedded
`<CR>`s that seem to mess up the parsing for `<Cmd>`, so I left them
as-is.
I also added a mapping of Space to no-op so that if I press space as my
leader and then don't press anything else quickly enough it won't
advance me a character.
There was also one mapping where I'd accidentally left off the closing
`>` on the end of the mapping, so I fixed that so that the mapping
works. Tested all the mappings to make sure everything's working as
expected.