Add some autocommands for filetype-specific mappings to Neovim config.
These are from the "Learn Vimscript the Hard Way" book. I originally
added these on the other branch for doing the exercises from the book,
and translated them into Lua for my Neovim config.
There's a bunch of other settings and mappings and autocommands and
things from the book, but these seem to be the most useful for right
now. I can always add more of them later if I want to.
- Comment out lines in different languages
- Fold the current tag in html file