Switch to using vim-plug for managing Vim plugins.
Obviously the situation is nicer in Neovim with lazy.nvim, but this is
still better than manually having the plugins installed as git submodules.
I might want to add another autocommand for running PlugInstall whenever
the plugins.vim file gets saved, but this works for now.
I also removed the taglist and OmniCppComplete plugins that I wasn't
using and that don't really work without ctags, which I haven't used for
a long time. If I decide to start using it again and want to set it up,
then I can easily add these plugins (or others that might work even better)
back into my config.
At this point I'm mainly just maintaining my Vim config for the few cases
in which I have to use vanilla Vim and Neovim isn't available, and so I
mainly just need it to be workable on a basic level. I'm not really going
to be playing with a bunch of things here like I will be doing in the
Neovim config. Making this switch to vim-plug is mainly to get away from
having git submodules in the repo that complicate cloning and branch
switching, etc.