Set the eol property for lazy-lock.json to lf in .gitattributes file.
I was seeing that when the lazy.nvim plugin touched the file it seemed
to cause it to show up as modified in git even though it was the same
except for line endings. Adding this .gitattributes file forces git to
always treat it as having LF endings and never convert it to CRLF
endings, which I think should make it so that lazy.nvim doesn't change
the line endings when it touches the file causing it to show up as
modified.