Fix lazy.nvim bootstrapping for older versions of Neovim.
I had thought that since the help said that vim.loop was deprecated and
that you should use vim.uv now that it was safe to just use vim.uv
instead of having the or in there, but on my Ubuntu WSL the version of
Neovim is a bit older and it's before the vim.uv functionality was
added. Adding in the or to use vim.loop as a backup fixes things.