Switch Vim config to use space for leader and add some mappings.
This is mainly just transferring the same thing I did with the leader in
my Neovim config over to the Vim config, and then adding some mappings
from doing the exercises for "Learn Vimscript the Hard Way" that I think
will be useful, and one from my Neovim config as well. I also added the
open previous buffer mapping to my Neovim config since it's handy for
when I close a window and realize that I actually wanted to keep it
open.