Fix some keybindings in VSCode.
I figured out a way to fix the alt keymaps as well as the window
navigation keymaps. I was able to add a keybindings.json file that I
then symlink to the right spot in my VSCode user directory. This might
cause some problems with settings syncing, but I'll deal with that if it
becomes a problem. For now I've got passthrough keybindings set up that
make things work for my keymaps.