Install telescope.nvim plugin.
There's probably a lot more things that I could explore with telescope
and make more mappings and things, but this is a good start. The idea
for the mapping to search over files in my Neovim config came from TJ
DeVries' video about telescope in his Advent of Neovim playlist.