Add some more keymaps/keybindings for VSCode.
Fixed up some of the existing ones to check for the neovim extension
being initialized, etc., and added mappings for easily toggling the
sidebar and panel, triggering quick open for files, opening the command
palette, etc. Some of what's needed to make shortcuts like this work
when there's no file open is adding passthrough keybindings. It's maybe
better to just define the keybindings directly instead of sending them
to Neovim and then having it turn right around and send a command back
to VSCode, but this is working for now, and it's nice that it keeps all
of the actual commands inside of the vscode-keymaps.lua file and doesn't
duplicate them in the keybindings.json file. We'll see if I come back to
this and decide I want to do it differently later.