Switch explorer opening to <Leader>E and work on mappings for editing configs.
I had originally had the mapping <Leader>e to open the file explorer, but
that meant that I couldn't have <Leader>en to edit my Neovim config or
<Leader>ev to edit my vimrc in Vim like is suggested in TJ DeVries' videos
and Learn Vimscript the Hard Way, respectively. I realized I could just as
easily make the file explorer toggle <Leader>E and then I could use the
suggested mappings for easily editing my configs. This also brings the
mapping sort of more in line with the default keybinding in VSCode for
toggling the explorer, Ctrl+Shift+E, just with my leader key instead of
Ctrl.
For VSCode, I made the mapping a bit different than the normal Telescope
mapping in Neovim, since things are a bit different with how it handles
folders and workspaces and whatnot. I have it just open the Neovim config
folder in a new VSCode window if it's not already open (though my next
change is going to simplify things because I discovered that I don't
actually need to do my own check for if the folder is already open). I can
then use my <Leader>ff mapping from within that new window to search for/
open files from within that folder.
I also added a mapping for doing the same thing, but with my whole .dotfiles
folder, since when I'm making edits to the config in VSCode I'll want to
be able to edit the keybindings.json file, which is outside of the nvim
config folder. This is now possible with <Leader>ed (for "edit dotfiles").

As part of making changes to my vscode mappings, I realized that I could
use a Lua function for setting the keymaps instead of having to use the
`lua` Vimscript command. I made a helper function to allow me to more
concisely write my mappings that just call to do VSCode actions.