Clean things up a bit from the last commit.
In my vscode mappings I figured out that I don't have to check myself for
if the folder I'm opening is already open in a VSCode window, so that
allowed me to simplify things quite a bit. I also did some renaming, made
the helper functions local, and created a local variable to store the
imported vscode module. I also switched the `open_folder` helper function
to use the same pattern as the other one where it will return a function
that can be passed directly to `vim.keymap.set` to simplify things.