Improve handling of Neovim keymaps for VSCode.
Most of this was generated using GitHub Copilot, with some manual adjustments.
I was tired of having to manually add passthrough keybindings for VSCode to
send the keypresses to Neovim when an editor wasn't open, and so I figured
out a way to automatically add the keybindings from the Neovim Lua config
itself when I define the original keymaps.
It's not perfect and a bit messy, but it's working for now. I'll likely
clean things up and refactor it at some point.
This also makes it so that there can be keybindings locally on a machine
that aren't added by the Neovim config, and those won't get overridden by
the custom keybindings being added.
I do need to test this on non-Windows machines, but I think it should work
as long as the paths to the keybindings.json file are correct.