Change the batch file to call powershell.exe instead of powershell.
This should get around the issue where in some environments there's a
powershell.bat wrapper on the path before the regular powershell.exe, and
that makes it so the bit at the end of the setupWin.ps1 script that pauses
before exiting doesn't work. This *should* fix that by forcing it to use
the real powershell.exe instead of the wrapper.