Update mappings in vsvimrc to match vimrc.
There's more mappings in vimrc that I'd like to add into vsvimrc, but
I'm not certain which ones will work within VsVim, so I just updated the
mappings that were already there to be the new keybindings.
Also updated a comment in vimrc that was referring to an old mapping.