Fix edit config file mapping in vsvimrc.
For some reason the same commands are working differently inside the mapping as opposed to executed
outright. Seems like a bug in or at least quirk of VsVim. Using `:tabedit` simply opens the file in
a new tab without replacing any open files or anything like that. I can live without the split.