Update modes for a bunch of keymaps across Vim, Neovim, and VsVim.
I realized that a bunch of the keymaps in my vimrc and vsvimrc were using just the standard `noremap`,
which applies to normal, visual, and operator-pending modes, and that doesn't necessarily make sense for
a lot of the mappings. Some work best only in normal mode and some are okay in normal and visual mode,
but none of them really make much sense in operator-pending mode. In fixing this I also realized that
in Neovim I had these same mappings just using normal mode for the most part, and so I also expanded
the mappings to visual mode there. It's a lot cleaner using the Lua API in Neovim since you can map the
desired set of modes explicitly, whereas in Vim if you want a combination of modes that isn't already
represented by a specific mapping command, then you have to duplicate the lines for those mappings for
each mode you want them in.