Add a keymap to dismiss GitHub Copilot NES with ESC in normal mode.
The vscode-neovim extension captures the ESC quickly enough in normal mode that it doesn't dismiss the Next Edit Suggestion if one is showing, which is really annoying. I was able to work around this by adding a keymap that maps ESC in normal mode (which usually is a no-op) to send the action to vscode for dismissing/hiding the inline suggestion.