Make a setup config file and cross-platform setup scripts.
I used GitHub Copilot rather heavily for this, and iterated a lot with it to find a good workable solution. I made some minor edits manually to get some things to work.
I think this should be a pretty good solution for me so that I can easily set up the symlinks and things that I want on different platforms. The psv config file should be fairly easy to keep up to date as the single source of truth for where we want symlinks to get created.