Neovim colorscheme fix for terminals without true color support.
When running Neovim in the default terminal on Mac OS, there isn't
proper true color support. Neovim detects this and sets or clears
`termguicolors` for you as appropriate based on the terminal
capabilities, but setting the Tokyo Night color scheme automatically
sets the `termguicolors` option for you, even if the terminal doesn't
have full color support. This results in things looking very weird.
The fix is to use an autocommand to set the color scheme once we get the
UIEnter event, but because of a race condition that doesn't seem to be
early enough all the time for the terminal probing to have happened, so
I used `vim.schedule` to delay the setting of the color scheme even
further to make sure that the terminal probing has finished and
`termguicolors` is set properly if it's supposed to be.
I had also been setting the `termguicolors` option manually before, and
so I removed that to let Neovim do its own determination of whether or
not to turn it on based on the terminal capabilities.